Allow to render documents to any paintable surfaces
For applications using the API, it is good if they can render the laid
out document to a preview area, or something. The document renderer has
been adapted to function in "headless mode", this is instead of
rendering the document to a managed pdf writer, it returns the layout
and then allow to use any painter to draw any item from the layout.